chore: clean up FormatString
