+++ journeyapps-evaluator/src/tools.ts
@@ -1,157 +1,9 @@
-// Unescape double closing braces to a single brace
-import { TokenExpression } from './token-expressions/TokenExpression';
-import { ConstantTokenExpression } from './token-expressions/ConstantTokenExpression';
 import { FunctionTokenExpression } from './token-expressions/FunctionTokenExpression';
 import { ShorthandTokenExpression } from './token-expressions/ShorthandTokenExpression';
 import { FormatShorthandTokenExpression } from './token-expressions/FormatShorthandTokenExpression';
 import { FormatString } from './FormatString';
 import { TypeInterface } from './TypeInterface';
 import { LegacyFunctionTokenExpression } from './token-expressions/LegacyFunctionTokenExpression';
-
-export function unescape(s: string) {
-  var start = 0;
-  var result = '';
-  var len = s.length;
-  while (true) {
-    var i = s.indexOf('}', start);
-    if (i == -1 || i == len - 1) {
-      result += s.substring(start);
-      break;
-    }
-    result += s.substring(start, i + 1);
-    // We assume that the character at i+1 is another right brace, but we don't do any checking.
-    start = i + 2;
-  }
-  return result;
-}
-
-export function parseEnclosingBraces(format: string) {
-  var i = format.indexOf('{');
-  if (i == -1) {
-    return null;
-  }
-
-  // We want to skip through these sections
-  // i.e. do not match { in a string, e.g. "{"
-  var SPECIAL_SECTIONS = ["'", '"'];
-
-  for (var k = i + 1; k < format.length; k++) {
-    var character = format[k];
-    if (SPECIAL_SECTIONS.indexOf(character) != -1) {
-      // This is the start of a string, jump to its end
-      var endChar = format.indexOf(character, k + 1);
-      if (endChar == -1) {
-        // Unless the end doesn't exist. Error out.
-        return null;
-      }
-      k = endChar;
-      continue;
-    }
-    if (character == '{') {
-      // Start of pair of inner braces,
-      // recursively parse them
-      var inner = parseEnclosingBraces(format.substring(k));
-      if (!inner) {
-        // Faulty inner, return null
-        return null;
-      }
-      k += inner.length;
-      continue;
-    }
-    if (character == '}') {
-      // Found closing part for current level of braces
-      // Return the length to the caller
-      return {
-        length: k - i
-      };
-    }
-  }
-  // Came to end of loop without a match. Faulty, return null
-  return null;
-}
-
-/**
- * Compile a format string expression into tokens.
- * @param {string} format string expression
- * @return {TokenExpression[]} compiled tokens
- */
-export function compile(format: string) {
-  let start = 0;
-
-  let tokens: TokenExpression[] = [];
-
-  let len = format.length;
-  while (true) {
-    const i = format.indexOf('{', start);
-    if (i < 0 || i == len - 1) {
-      // end of string - everything is normal text
-      tokens.push(new ConstantTokenExpression(unescape(format.substring(start)), start));
-      break;
-    }
-    // normal text in the gaps between curly braces
-    tokens.push(new ConstantTokenExpression(unescape(format.substring(start, i)), start));
-    if (format[i + 1] == '{') {
-      // Double left brace - escape and continue
-      tokens.push(new ConstantTokenExpression('{', start));
-      start = i + 2;
-      continue;
-    }
-
-    const parsedBraces = parseEnclosingBraces(format.substring(i));
-    if (!parsedBraces) {
-      // Brace pair faulty (no closing brace), return as a constant
-      tokens.push(new ConstantTokenExpression(format.substring(i), start));
-      break;
-    }
-
-    // Next start is at the end of the currently parsed brace pair
-    start = i + parsedBraces.length + 1;
-
-    // `spec` is everything between the curly braces "{" and "}".
-    const spec = format.substring(i + 1, i + parsedBraces.length);
-
-    // test for function token prefix
-    if (spec.trim().indexOf(FunctionTokenExpression.PREFIX) === 0) {
-      // function token because the function name has "$:" as prefix (leading whitespace is ignored)
-      tokens.push(new FunctionTokenExpression(spec, i));
-    } else {
-      // shorthand token
-      const colon = spec.indexOf(':');
-      if (colon == -1) {
-        tokens.push(new ShorthandTokenExpression(spec, i));
-      } else {
-        tokens.push(new FormatShorthandTokenExpression(spec.substring(0, colon), spec.substring(colon + 1), i));
-      }
-    }
-  }
-
-  // concatenate any neighbouring constant token expressions
-  let result: TokenExpression[] = [];
-  let last: ConstantTokenExpression = null;
-  for (var j = 0; j < tokens.length; j++) {
-    var token = tokens[j];
-    if (token instanceof ConstantTokenExpression) {
-      if (last == null) {
-        if (token.expression.length > 0) {
-          last = token;
-        }
-      } else {
-        last = last.concat(token);
-      }
-    } else {
-      if (last != null) {
-        result.push(last);
-        last = null;
-      }
-      result.push(token);
-    }
-  }
-  if (last != null) {
-    result.push(last);
-  }
-
-  return result;
-}
 
 export function getObjectType(parent: any, name: string) {
   var variable = parent.getAttribute(name);
@@ -289,4 +141,4 @@ export async function formatValueAsync(value: any, type: TypeInterface, format: 
 }
 
 // Expose internal functions for tests
-export { compile as _compile, deepMerge as _deepMerge };
+export { deepMerge as _deepMerge };
